+++ controllers/html/dashboard-routes.js
@@ -1,7 +1,7 @@
 const router = require('express').Router();
-const sequelize = require('../config/connection');
-const { User, Deck, Deck_Components } = require('../models');
-const withAuth = require('../utils/auth');
+const sequelize = require('../../config/connection');
+const { User, Deck, Deck_Components } = require('../../models');
+const withAuth = require('../../utils/auth');
 
 router.get('/', (req, res) => {
 	// console.log(req.session);
