+++ controllers/index.js
@@ -1,11 +1,9 @@
 const router = require('express').Router();
-const homeRoutes = require('./home-routes.js');
 const apiRoutes = require('./api');
-const dashboardRoutes = require('./dashboard-routes.js');
+const htmlRoutes = require('./html');
 
-router.use('/dashboard', dashboardRoutes);
 router.use('/api', apiRoutes);
-router.use('/', homeRoutes);
+router.use('/', htmlRoutes);
 
 router.use((req, res) => {
   res.status(404).end();
